+++ app/core/src/main/java/com/teamg/tourdeshot/core/mapper/ProductMapperImpl.java
@@ -4,7 +4,6 @@ import com.teamg.tourdeshot.core.api.local.domain.ProductDTO;
 import com.teamg.tourdeshot.core.model.Product;
 import org.springframework.stereotype.Component;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 @Component
@@ -12,13 +11,13 @@ public class ProductMapperImpl implements ProductMapper {
 
     @Override
     public ProductDTO toProductDTO(Product product) {
-        if(Objects.isNull(product))
+        if (Objects.isNull(product))
             return null;
         return ProductDTO.builder()
                 .productId(product.getProductId())
                 .name(product.getName())
                 .description(product.getDescription())
-                .price(BigDecimal.valueOf(product.getPrice()))
+                .price(product.getPrice())
                 .ingredients(product.getIngredients())
                 .build();
     }
